Rename PascalCase page files to kebab-case
Renames all scaffold page files to follow the kebab-case convention
introduced in STYLE_GUIDE.md and updates imports in routes.tsx.
