+++ apps/web/src/pages/library-detail.tsx
